agent/pyproject.toml
[project]
name = "relate"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-openai",
    "langgraph>=0.5.0",
    "langgraph-cli[inmem]",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    "openai>=1.93.0",
    "rich>=14.0.0",
    "ruff>=0.8.0",
    "langsmith>=0.4.4",
    "matplotlib>=3.7.0",
    "pandas>=2.3.1",
    "langmem>=0.0.28",
]

[project.scripts]
relate = "src.cli:main"
relate-agent = "app.main:main"

[tool.setuptools]
packages = ["app"]

[tool.uv]
package = true
